backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
phonenumbers==8.13.27

# AI/LLM
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.26

# Integrations
twilio==8.13.0
httpx==0.26.0

# Background Jobs
celery==5.3.6
celery[redis]==5.3.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
